requirements.txt
celery
python-dotenv
pymongo
curl_cffi
selectolax